+++ R/docx2pdf.R
@@ -61,10 +61,10 @@ docx2pdf <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", input)) 
   }
 }
 
+
 #' @importFrom locatexec is_osx
 #' @importFrom processx run
 docx2pdf_osx <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", input)){
-
   if (!is_osx()) {
     stop("docx2pdf_osx() should only be used on 'macOS' systems.", call. = FALSE)
   }
@@ -101,6 +101,7 @@ docx2pdf_osx <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", inpu
   output
 }
 
+
 #' @importFrom locatexec is_windows
 docx2pdf_win <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", input)){
 
@@ -112,6 +113,8 @@ docx2pdf_win <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", inpu
     stop("input does not exist", call. = FALSE)
   }
 
+  stop_on_wrong_ps_exec_policy()
+
   input <- absolute_path(input)
   output <- absolute_path(output)
 
@@ -120,8 +123,7 @@ docx2pdf_win <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", inpu
 
   output_name <- file.path(default_root, gsub("\\.(docx|doc|rtf)$", ".pdf", basename(input)))
 
-  script_sourcefile <- system.file(
-    package = "doconv", "scripts", "powershell", "docx2pdf.ps1")
+  script_sourcefile <- system.file(package = "doconv", "scripts", "powershell", "docx2pdf.ps1")
   script_path <- tempfile(fileext = ".ps1")
   script_str <- readLines(script_sourcefile, encoding = "UTF-8")
   script_str[1] <- sprintf(script_str[1], input)
@@ -140,9 +142,8 @@ docx2pdf_win <- function(input, output = gsub("\\.(docx|doc|rtf)$", ".pdf", inpu
   if(!success) stop("could not convert ", input, res$stderr, call. = FALSE)
 
   output
-
-
 }
+
 
 #' @export
 #' @title Update docx fields
@@ -178,6 +179,7 @@ docx_update <- function(input) {
   invisible(x)
 }
 
+
 docx_update_osx <- function(input){
 
   if (!is_osx()) {
@@ -203,6 +205,8 @@ docx_update_osx <- function(input){
 
   success
 }
+
+
 docx_update_win <- function(input){
 
   if (!is_windows()) {
@@ -212,6 +216,8 @@ docx_update_win <- function(input){
   if (!file.exists(input)) {
     stop("input does not exist", call. = FALSE)
   }
+
+  stop_on_wrong_ps_exec_policy()
 
   input <- absolute_path(input)
 
@@ -231,10 +237,6 @@ docx_update_win <- function(input){
 
   success
 }
-
-
-
-
 
 
 #' @export
@@ -290,6 +292,7 @@ pptx2pdf <- function(input, output = gsub("\\.pptx$", ".pdf", input)) {
   }
 }
 
+
 #' @importFrom locatexec is_osx
 #' @importFrom processx run
 pptx2pdf_osx <- function(input, output = gsub("\\.pptx$", ".pdf", input)){
@@ -332,6 +335,7 @@ pptx2pdf_osx <- function(input, output = gsub("\\.pptx$", ".pdf", input)){
   output
 }
 
+
 #' @importFrom locatexec is_windows
 pptx2pdf_win <- function(input, output = gsub("\\.pptx$", ".pdf", input)){
 
@@ -342,6 +346,8 @@ pptx2pdf_win <- function(input, output = gsub("\\.pptx$", ".pdf", input)){
   if (!file.exists(input)) {
     stop("input does not exist", call. = FALSE)
   }
+
+  stop_on_wrong_ps_exec_policy()
 
   input <- absolute_path(input)
   output <- absolute_path(output)
@@ -359,8 +365,5 @@ pptx2pdf_win <- function(input, output = gsub("\\.pptx$", ".pdf", input)){
 
   if(!success) stop("could not convert ", input, call. = FALSE)
 
-  success
-
-
+  output
 }
-