+++ NEWS.md
@@ -1,5 +1,9 @@
 # doconv 0.3.2
 
+## Features
+
+- add support for 'RTF'.
+
 ## Issues
 
 - For 'Windows' users, figure resolution should now remain the same