+++ NEWS.md
@@ -1,3 +1,10 @@
+# doconv 0.3.3 (dev version)
+
+## Issues
+
+- Stop if PowerShell (PS) execution strategy does not allow running scripts and advise user. 
+  PS scripts are required for certain actions on Windows (#2). 
+
 # doconv 0.3.2
 
 ## Features